fix: build script fix

+++ build.sh
@@ -8,7 +8,8 @@ cp -r dist/* ../public/markdown-viewer/
 cd ..
 go build
 mkdir -p dist
-cp simple-server dist/
+[ -f simple-server ] && cp simple-server dist/
+[ -f simple-server.exe ] && cp simple-server.exe dist/
 [ -f config.yaml ] && cp config.yaml dist/
-cp config.yaml.example dist/
+[ -f config.yaml.example ] && cp config.yaml.example dist/
 cp -r public dist/
